feat: set variables (#40)
Allows the user to set the value of variables within the variable/hover components. This uses the `setExpression` call if available and falls back to the `setVariable` call which is described as the desired behaviour in the specification. If a debug server doesn't support `setExpression` the hover window cannot set the value of a variable in the top level as we have no way of knowing the required parent `variableReference`