fix(variables): handle changing child var ref
Variable references can change at any time. If a variable was hidden and
re-opened with one of its child var refs changed which was previously
opened, the render would always be invalidated due to state not
containing the variables. This change makes sure that the expanded
variables are always monitored.

See #47