Implement separate *NC control highlight groups
For each of the control highlight groups there is not a corresponding
highlight group with the same name plus 'NC' as suffix. When entering or
exiting the window which carries the support groups we reset the
'winbar' option such that the highlight groups are either normal or
'*NC'.

The reason for this is that the user might have a different background
for the highlight groups 'WinBar' and 'WinBarNC'. Whenever the user
exits or enters a window, the editor will flip between these two groups,
so we have to flip our custom highlight groups accordingly as well.