fix(jump_to_frame): make sure that source.path exists before jumping to it
When you try to Open a Frame from DAPStacks window,
if Adapter reported source.path for said Frame
but the path does not actually exist in the file system,
client_lib.jump_to_frame() opens empty buffer for it
and jumps to non-existing line (causing exception)
INSTEAD OF reporting an error and early returning.
This commit fixes said problem.